+++ Nos/Controller/NoteWarningController.swift
@@ -25,14 +25,20 @@ extension Publisher {
     }
 }
 
-@Observable class NoteWarningController {
+@Observable @MainActor class NoteWarningController {
     
-    var showWarning = false
-    var userHidWarning = false {
-        didSet {
-            computeShowWarning()
+    var showWarning: Bool {
+        if userHidWarning {
+            return false
+        } else if showReportWarnings && (!noteReports.isEmpty || !authorReports.isEmpty) {
+            return true
+        } else if shouldHideOutOfNetwork && showOutOfNetworkWarning && outOfNetwork {
+            return true
+        } else {
+            return false
         }
     }
+    var userHidWarning = false 
     var outOfNetwork = false
     
     var note: Event? {
@@ -48,16 +54,8 @@ extension Publisher {
     @ObservationIgnored @Dependency(\.persistenceController) var persistenceController
     @ObservationIgnored @Dependency(\.userDefaults) var userDefaults
     
-    var noteReports = [Event]() {
-        didSet {
-            computeShowWarning()
-        }
-    }
-    var authorReports = [Event]() {
-        didSet {
-            computeShowWarning()
-        }
-    }
+    var noteReports = [Event]() 
+    var authorReports = [Event]() 
     private var cancellables = [AnyCancellable]()
     private var noteReportsWatcher: NSFetchedResultsController<Event>?
     private var authorReportsWatcher: NSFetchedResultsController<Event>?
@@ -96,14 +94,12 @@ extension Publisher {
                 FetchedResultsControllerPublisher(fetchedResultsController: noteReportsWatcher)
                     .publisher
                     .asyncMap { (events: [Event]?) -> [Event] in
-                        await MainActor.run {
-                            var eventsFromFollowedAuthors = [Event]()
-                            for event in events ?? [] where
-                                self.currentUser.socialGraph.follows(event.author?.hexadecimalPublicKey) {
-                                    eventsFromFollowedAuthors.append(event)
-                            }
-                            return eventsFromFollowedAuthors
+                        var eventsFromFollowedAuthors = [Event]()
+                        for event in events ?? [] where
+                        await self.currentUser.socialGraph.follows(event.author?.hexadecimalPublicKey) {
+                            eventsFromFollowedAuthors.append(event)
                         }
+                        return eventsFromFollowedAuthors
                     } 
                     .receive(on: DispatchQueue.main)
                     .sink(receiveValue: { [weak self] followedReports in
@@ -123,14 +119,12 @@ extension Publisher {
                     FetchedResultsControllerPublisher(fetchedResultsController: authorReportsWatcher)
                         .publisher
                         .asyncMap { (events: [Event]?) -> [Event] in
-                            await MainActor.run {
-                                var eventsFromFollowedAuthors = [Event]()
-                                for event in events ?? [] where 
-                                self.currentUser.socialGraph.follows(event.author?.hexadecimalPublicKey) {
-                                    eventsFromFollowedAuthors.append(event)
-                                }
-                                return eventsFromFollowedAuthors
+                            var eventsFromFollowedAuthors = [Event]()
+                            for event in events ?? [] where 
+                            await self.currentUser.socialGraph.follows(event.author?.hexadecimalPublicKey) {
+                                eventsFromFollowedAuthors.append(event)
                             }
+                            return eventsFromFollowedAuthors
                         } 
                         .receive(on: DispatchQueue.main)
                         .sink(receiveValue: { [weak self] followedReports in
@@ -145,7 +139,7 @@ extension Publisher {
             .sink { note in
                 Task { @MainActor in 
                     if let authorKey = note?.author?.hexadecimalPublicKey {
-                        self.outOfNetwork = !self.currentUser.socialGraph.contains(authorKey)
+                        self.outOfNetwork = await !self.currentUser.socialGraph.isInNetwork(authorKey)
                     } else {
                         self.outOfNetwork = false
                     }
@@ -168,17 +162,5 @@ extension Publisher {
                 showOutOfNetworkWarning = userDefaults.object(forKey: showOutOfNetworkWarningKey) as? Bool ?? true
             }
             .store(in: &cancellables)
-    }
-    
-    func computeShowWarning() {
-        if userHidWarning {
-            self.showWarning = false
-        } else if showReportWarnings && !(noteReports.isEmpty || !authorReports.isEmpty) {
-            self.showWarning = true
-        } else if shouldHideOutOfNetwork && showOutOfNetworkWarning && outOfNetwork {
-            self.showWarning = true
-        } else {
-            self.showWarning = false
-        }
     }
 }