+++ Nos/Views/AuthorStoryView.swift
@@ -98,7 +98,7 @@ struct AuthorStoryView: View {
                             RoundedRectangle(cornerRadius: 21)
                                 .frame(maxWidth: .infinity, maxHeight: 3)
                                 .cornerRadius(7)
-                                .foregroundColor(note.isEqual(selectedNote) == true ? .accent : .secondaryText)
+                                .foregroundColor(note.isEqual(selectedNote) == true ? .accent : .secondaryTxt)
                                 .padding(.bottom, 5)
                                 .padding(.top, 15)
                         }
@@ -112,20 +112,23 @@ struct AuthorStoryView: View {
                     HStack(alignment: .center) {
                         AuthorLabel(author: author)
                             .padding(0)
-                        if let expirationTime = selectedNote?.expirationDate?.distanceFromNowString() {
+                        if selectedNote?.kind == EventKind.repost.rawValue {
+                            Image.repostSymbol
+                        } 
+                        if let expirationTime = selectedNote?.expirationDate?.distanceString() {
                             Image.disappearingMessages
                                 .resizable()
-                                .foregroundColor(.secondaryText)
+                                .foregroundColor(.secondaryTxt)
                                 .frame(width: 25, height: 25)
                             Text(expirationTime)
                                 .lineLimit(1)
                                 .font(.body)
-                                .foregroundColor(.secondaryText)
-                        } else if let elapsedTime = selectedNote?.createdAt?.distanceFromNowString() {
+                                .foregroundColor(.secondaryTxt)
+                        } else if let elapsedTime = selectedNote?.createdAt?.distanceString() {
                             Text(elapsedTime)
                                 .lineLimit(1)
                                 .font(.body)
-                                .foregroundColor(.secondaryText)
+                                .foregroundColor(.secondaryTxt)
                         }
                         Spacer()
                         if let selectedNote {
@@ -214,11 +217,13 @@ fileprivate struct BottomOverlay: View {
                 } label: {
                     Text(replies)
                         .font(.subheadline)
-                        .foregroundColor(Color.secondaryText)
+                        .foregroundColor(Color.secondaryTxt)
                 }
             }
 
             Spacer()
+
+            RepostButton(note: note)
 
             LikeButton(note: note)