+++ Nos/Views/UNS/UNSWizardPhoneView.swift
@@ -38,7 +38,7 @@ struct UNSWizardPhoneView: View {
                     Text(.registrationDescription)
                         .fontWeight(.medium)
                         .multilineTextAlignment(.center)
-                        .foregroundColor(.secondaryText)
+                        .foregroundColor(.secondaryTxt)
                         .padding(.vertical, 17)
                         .padding(.horizontal, 20)
                         .shadow(radius: 1, y: 1)